% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfunc_minizinc.R
\name{add_solvers}
\alias{add_solvers}
\title{add solvers for the model (uses can be extended later )}
\usage{
add_solvers(include_name, append_bool = TRUE)
}
\arguments{
\item{include_name}{string containing the name of the solver to be used.}

\item{append_bool}{bool to tell if it is required to open the file in append mode. TRUE by default
currently only a few solvers like Geocode can be used, the code will be modified later}
}
\description{
add solvers for the model (uses can be extended later )
}
