% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable.R
\name{array_checks}
\alias{array_checks}
\title{array checks}
\description{
this class contains the various checks required for array values
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{array_checks$new()}}
\item \href{#method-clone}{\code{array_checks$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
this constructor contains the checks for array values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{array_checks$new(
  array_kind,
  array_type,
  array_value = NULL,
  array_index = NULL,
  array_expr = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{array_kind}}{parameter or decision}

\item{\code{array_type}}{int, float, bool, set or enum}

\item{\code{array_value}}{value of the array (only for parameters)}

\item{\code{array_index}}{index of the array}

\item{\code{array_expr}}{expression to be given to the array}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{array_checks$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
