% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{AssignItem}
\alias{AssignItem}
\title{Assignment Items}
\description{
assignments in MiniZinc
}
\section{Super class}{
\code{\link[rminizinc:Item]{rminizinc::Item}} -> \code{AssignItem}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{.decl}}{associated declaration}

\item{\code{.e}}{value to be assigned}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{.decl}}{associated declaration}

\item{\code{.e}}{value to be assigned}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AssignItem$new()}}
\item \href{#method-id}{\code{AssignItem$id()}}
\item \href{#method-getV}{\code{AssignItem$getV()}}
\item \href{#method-setV}{\code{AssignItem$setV()}}
\item \href{#method-getDecl}{\code{AssignItem$getDecl()}}
\item \href{#method-setDecl}{\code{AssignItem$setDecl()}}
\item \href{#method-c_str}{\code{AssignItem$c_str()}}
\item \href{#method-clone}{\code{AssignItem$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AssignItem$new(decl, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{decl}}{declaration associated with assignment}

\item{\code{value}}{expression to be assigned}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-id"></a>}}
\if{latex}{\out{\hypertarget{method-id}{}}}
\subsection{Method \code{id()}}{
get the name of assigned variable
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AssignItem$id()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getV"></a>}}
\if{latex}{\out{\hypertarget{method-getV}{}}}
\subsection{Method \code{getV()}}{
get the value
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AssignItem$getV()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setV"></a>}}
\if{latex}{\out{\hypertarget{method-setV}{}}}
\subsection{Method \code{setV()}}{
set the value
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AssignItem$setV(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{value/expression to be set}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getDecl"></a>}}
\if{latex}{\out{\hypertarget{method-getDecl}{}}}
\subsection{Method \code{getDecl()}}{
get the associated declaration
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AssignItem$getDecl()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setDecl"></a>}}
\if{latex}{\out{\hypertarget{method-setDecl}{}}}
\subsection{Method \code{setDecl()}}{
set the associated declaration
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AssignItem$setDecl(decl)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{decl}}{declaration to be set}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-c_str"></a>}}
\if{latex}{\out{\hypertarget{method-c_str}{}}}
\subsection{Method \code{c_str()}}{
get the MiniZinc representation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AssignItem$c_str()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AssignItem$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
