% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommonMZNProblems.R
\name{assignment}
\alias{assignment}
\title{assignment problem}
\usage{
assignment(n, m, cost)
}
\arguments{
\item{n}{number of agents}

\item{m}{number of tasks}

\item{cost}{m x n 2D array where each row corresponds to the cost of each task for that agent.
(to be provided as 1-D vector)}
}
\description{
Solve an assignment problem (Goal is to minimize the cost)
}
