% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommonMZNProblems.R
\name{assignment_2}
\alias{assignment_2}
\title{assignment problem 2}
\usage{
assignment_2(rows, cols, cost)
}
\arguments{
\item{rows}{number of columns}

\item{cols}{number of tasks}

\item{cost}{cost matrix (to be provided as 1-D vector)}
}
\description{
Solve an assignment problem 
Winston "Operations Research", page 398, swimming team example
Model created by Hakan Kjellerstrand(hakank(at)bonetmail.com)
See : http://www.hakank.org/minizinc/assignment2.mzn
}
