% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Expression.R
\name{Binop}
\alias{Binop}
\title{Binop class}
\description{
create a binary operator expression
}
\section{Super class}{
\code{\link[rminizinc:Expression]{rminizinc::Expression}} -> \code{Binop}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{.lhs_exp}}{the left hand side expression}

\item{\code{.rhs_exp}}{the right hand side expression}

\item{\code{.op}}{the operator}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{.lhs_exp}}{the left hand side expression}

\item{\code{.rhs_exp}}{the right hand side expression}

\item{\code{.op}}{the operator}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Binop$new()}}
\item \href{#method-lhs}{\code{Binop$lhs()}}
\item \href{#method-rhs}{\code{Binop$rhs()}}
\item \href{#method-op}{\code{Binop$op()}}
\item \href{#method-clone}{\code{Binop$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Binop$new(lhs_expression, binop, rhs_expression)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lhs_expression}}{the left hand side expression}

\item{\code{binop}}{the binary operator to be used}

\item{\code{rhs_expression}}{the right hand side expression}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-lhs"></a>}}
\if{latex}{\out{\hypertarget{method-lhs}{}}}
\subsection{Method \code{lhs()}}{
return the lhs expression
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Binop$lhs()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rhs"></a>}}
\if{latex}{\out{\hypertarget{method-rhs}{}}}
\subsection{Method \code{rhs()}}{
return the rhs expression
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Binop$rhs()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-op"></a>}}
\if{latex}{\out{\hypertarget{method-op}{}}}
\subsection{Method \code{op()}}{
return the operator
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Binop$op()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Binop$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
