% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeclFunctions.R
\name{BoolSetDecl}
\alias{BoolSetDecl}
\title{set of bool declaration}
\usage{
BoolSetDecl(name, kind, value = NULL)
}
\arguments{
\item{name}{variable/parameter name}

\item{kind}{"var" or "par"}

\item{value}{value of the set (or NULL)}
}
\description{
Declare a new set of bool
}
