% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfunc_minizinc.R
\name{check_results}
\alias{check_results}
\title{Function to check the solution}
\usage{
check_results(
  path_of_model_mzn,
  path_of_data_dzn,
  path_of_model_mzc_mzn,
  data_file_bool = FALSE,
  compile_bool = FALSE,
  path_of_model_mzc = " ",
  compiled = FALSE
)
}
\arguments{
\item{path_of_model_mzn}{the path of the model.mzn file}

\item{path_of_data_dzn}{the path of the data.dzn file}

\item{path_of_model_mzc_mzn}{the path of model.mzc.mzn file}

\item{data_file_bool}{the bool to determine if you want to use data file to initialize the parameters}

\item{compile_bool}{the bool to determine if you wish to compiler the checker file}

\item{path_of_model_mzc}{the path of the compiled checker file model.mzc}

\item{compiled}{the bool to determine if the checker file is alrady compiled}
}
\description{
Function to check the solution
}
