% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constraint.R
\name{ConstraintItem}
\alias{ConstraintItem}
\title{Constraint Class}
\description{
Constraint is a class to describe Minizinc constraints on decision variables.
}
\section{Super class}{
\code{\link[rminizinc:Item]{rminizinc::Item}} -> \code{ConstraintItem}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{.expression}}{the constraint expression}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{.expression}}{the constraint expression}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ConstraintItem$new()}}
\item \href{#method-e}{\code{ConstraintItem$e()}}
\item \href{#method-c_str}{\code{ConstraintItem$c_str()}}
\item \href{#method-clone}{\code{ConstraintItem$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of Constraint class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConstraintItem$new(expression)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{expression}}{The expression for the constraint}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-e"></a>}}
\if{latex}{\out{\hypertarget{method-e}{}}}
\subsection{Method \code{e()}}{
return the constraint expression
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConstraintItem$e()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-c_str"></a>}}
\if{latex}{\out{\hypertarget{method-c_str}{}}}
\subsection{Method \code{c_str()}}{
serialize to MiniZinc syntax
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConstraintItem$c_str()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConstraintItem$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
