% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfunc_minizinc.R
\name{create_data_dzn}
\alias{create_data_dzn}
\title{create data file}
\usage{
create_data_dzn(input_var_names, input_var_values, path)
}
\arguments{
\item{input_var_names}{list of names of the input variables}

\item{input_var_values}{list of values of input variables}

\item{path}{the path of data dzn}
}
\description{
create data file
}
