% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfunc_minizinc.R
\name{create_decision_vars}
\alias{create_decision_vars}
\title{create the decision variables}
\usage{
create_decision_vars(
  single_dec_var_names,
  single_dec_var_ranges,
  dec_var_collection_names = c(),
  dec_collection_types = c(),
  dec_var_collection_types = c(),
  dec_var_array_ranges = c(),
  append_bool = TRUE,
  path
)
}
\arguments{
\item{dec_var_collection_names}{list of names of collection decision variables(array or set)}

\item{dec_collection_types}{list of types of collection.Possible types are "array" or "set".}

\item{dec_var_collection_types}{list of data types of collection of decision variables.}

\item{append_bool}{bool to tell if it is required to open the file in append mode. TRUE by default}

\item{path}{the path for model mzn}

\item{dec_var_names}{list of names of single decision variables}

\item{dec_var_ranges}{list of ranges of the single decision variables. Typical range should look like c(2,10).}

\item{array_var_ranges}{list of array ranges in order of  the array variables used.typical element in the list
should look like c(lower_limit,upper_limit), for eg. c(2,10) for 2D array and single numbers for eg 7 for
1D array .}
}
\description{
create the decision variables
}
