% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfunc_minizinc.R
\name{create_vars}
\alias{create_vars}
\title{create the variables}
\usage{
create_vars(
  single_var_names,
  single_var_types,
  collection_var_names = c(),
  collection_types = c(),
  collection_var_types = c(),
  array_var_ranges = c(),
  append_bool = FALSE,
  data_file_bool = TRUE,
  single_var_values = c(),
  collection_var_values = c(),
  path
)
}
\arguments{
\item{single_var_names}{list of names of basic/single variables/parameters(not collection)}

\item{single_var_types}{list of data types of basic/single variables.Possible types are "int"
"float", "string" and "bool".}

\item{collection_var_names}{list of names of collection variables/parameters(array or set)}

\item{collection_types}{list of types of collection.Possible types are "array" or
"set".}

\item{collection_var_types}{list of data types of collection variables/parameters.}

\item{array_var_ranges}{list of array ranges in order of  the array variables used.typical element in the list
should look like c(lower_limit,upper_limit), for eg. c(2,10) for 2D array and single numbers for eg 7 for
1D array .}

\item{append_bool}{bool to tell if it is required to open the file in append mode. FALSE by default}

\item{data_file_bool}{bool to determine if the user wants to initialize the parameters using a data file}

\item{single_var_values}{list of values of the single variables/parameters.}

\item{collection_var_values}{list of values of the collection variables.}

\item{path}{the path of the model.mzn file}
}
\description{
create the variables
}
