% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{deleteItem}
\alias{deleteItem}
\title{Delete an item from a MiniZinc model}
\usage{
deleteItem(
  itemNo,
  modelString = "",
  mznpath = "",
  modelStringName = "deleteItem.mzn",
  updateMZN = FALSE
)
}
\arguments{
\item{itemNo}{the item number of the item to be deleted.}

\item{modelString}{string representation of the MiniZinc model.}

\item{mznpath}{the path of model mzn.}

\item{modelStringName}{the name of model string.}

\item{updateMZN}{bool to specify if the new model should be written to given file.}
}
\description{
deletes an item from MiniZinc based on its item
number that is obtained from mzn_parse() function.
}
