% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_expression.R
\name{Expression}
\alias{Expression}
\title{function to generate an expression}
\description{
This class can be used to store an expression. A simple expression consists of variables and
operators. The data structure to be chosen for storing expressions is yet to be decided.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Expression$new()}}
\item \href{#method-clone}{\code{Expression$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
constructor of an abstract class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Expression$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Expression$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
