% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iter_delete.R
\name{expression_delete}
\alias{expression_delete}
\title{delete an expression(Under Development)}
\usage{
expression_delete(obj)
}
\arguments{
\item{obj}{object to delete}
}
\description{
delete the object everywhere from the MiniZinc model
}
