% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expressionDelete}
\alias{expressionDelete}
\title{delete an expression}
\usage{
expressionDelete(classNm, model)
}
\arguments{
\item{classNm}{class of the object to delete}

\item{model}{model to delete the object from}
}
\description{
Delete the object everywhere from the
MiniZinc model
}
