% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{filetoString}
\alias{filetoString}
\title{convert filename to string}
\usage{
filetoString(filepath)
}
\arguments{
\item{filepath}{the path of mzn file.}
}
\description{
this function reads a file from a given file path
and reads it.
}
