% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Value.R
\name{FloatSetVal}
\alias{FloatSetVal}
\title{Float set value}
\description{
float set range in MiniZinc
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{.min}}{minimum FloatVal}

\item{\code{.max}}{maximum FloatVal}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{.min}}{minimum FloatVal}

\item{\code{.max}}{maximum FloatVal}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FloatSetVal$new()}}
\item \href{#method-getMin}{\code{FloatSetVal$getMin()}}
\item \href{#method-setMin}{\code{FloatSetVal$setMin()}}
\item \href{#method-getMax}{\code{FloatSetVal$getMax()}}
\item \href{#method-setMax}{\code{FloatSetVal$setMax()}}
\item \href{#method-clone}{\code{FloatSetVal$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FloatSetVal$new(fmin, fmax)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fmin}}{the minimum FloatVal}

\item{\code{fmax}}{the maximum FloatVal}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getMin"></a>}}
\if{latex}{\out{\hypertarget{method-getMin}{}}}
\subsection{Method \code{getMin()}}{
get the minimum float value
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FloatSetVal$getMin()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setMin"></a>}}
\if{latex}{\out{\hypertarget{method-setMin}{}}}
\subsection{Method \code{setMin()}}{
set the minimum float value
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FloatSetVal$setMin(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{float value to be set}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getMax"></a>}}
\if{latex}{\out{\hypertarget{method-getMax}{}}}
\subsection{Method \code{getMax()}}{
get the maximum float value
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FloatSetVal$getMax()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setMax"></a>}}
\if{latex}{\out{\hypertarget{method-setMax}{}}}
\subsection{Method \code{setMax()}}{
set the maximum float value
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FloatSetVal$setMax(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{float value to be set}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FloatSetVal$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
