% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_expr.R
\name{generate_expr}
\alias{generate_expr}
\title{function to generate an expression}
\usage{
generate_expr(LHS, operator, RHS)
}
\arguments{
\item{LHS}{matrix}

\item{operator}{connecting operator}

\item{RHS}{matrix}
}
\description{
function to generate an expression
}
