% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Expression.R
\name{Generator}
\alias{Generator}
\title{Generator}
\description{
create a generator in MiniZinc
}
\section{Super class}{
\code{\link[rminizinc:Expression]{rminizinc::Expression}} -> \code{Generator}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{.decls}}{variable declarations}

\item{\code{.in}}{in expression}

\item{\code{where}}{where expression}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{.decls}}{variable declarations}

\item{\code{.in}}{in expression}

\item{\code{where}}{where expression}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Generator$new()}}
\item \href{#method-In}{\code{Generator$In()}}
\item \href{#method-where}{\code{Generator$where()}}
\item \href{#method-decl}{\code{Generator$decl()}}
\item \href{#method-c_str}{\code{Generator$c_str()}}
\item \href{#method-clone}{\code{Generator$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Generator$new(decls, IN = NULL, where = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{decls}}{list of variable declarations}

\item{\code{IN}}{the in expression of generator}

\item{\code{where}}{the where expression of generator}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-In"></a>}}
\if{latex}{\out{\hypertarget{method-In}{}}}
\subsection{Method \code{In()}}{
get the in expression
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Generator$In()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-where"></a>}}
\if{latex}{\out{\hypertarget{method-where}{}}}
\subsection{Method \code{where()}}{
get the where expression
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Generator$where()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-decl"></a>}}
\if{latex}{\out{\hypertarget{method-decl}{}}}
\subsection{Method \code{decl()}}{
get the ith declaration
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Generator$decl(i)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{i}}{index}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-c_str"></a>}}
\if{latex}{\out{\hypertarget{method-c_str}{}}}
\subsection{Method \code{c_str()}}{
get the MiniZinc representation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Generator$c_str()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Generator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
