% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_expression.R
\name{get_expression}
\alias{get_expression}
\title{function to generate an expression}
\description{
This class can be used to store an expression. A simple expression consists of variables and
operators. The data structure to be chosen for storing expressions is yet to be decided.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{expr}}{the generated expression from the inputs}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{expr}}{the generated expression from the inputs}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{get_expression$new()}}
\item \href{#method-get_expr}{\code{get_expression$get_expr()}}
\item \href{#method-clone}{\code{get_expression$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
the function to store an expression
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{get_expression$new(variables, arithmetic_operator = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{variables}}{the variables involved in the expression in order}

\item{\code{arithmetic_operator}}{the arithmetic operators involved in the function}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_expr"></a>}}
\if{latex}{\out{\hypertarget{method-get_expr}{}}}
\subsection{Method \code{get_expr()}}{
Returns the required expression
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{get_expression$get_expr()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{get_expression$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
