% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_missing_pars}
\alias{get_missing_pars}
\title{get missing parameters}
\usage{
get_missing_pars(
  modelString = "",
  mznpath = "",
  modelStringName = "missing_pars.h",
  includePath = NULL
)
}
\arguments{
\item{modelString}{the string representation of a MiniZinc model}

\item{mznpath}{the path of the MiniZinc model mzn file}

\item{modelStringName}{the custom name of the mzn string}

\item{includePath}{path of the included mzn in the model if it exists.}
}
\description{
get the names of missing parameters in a model
}
