% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mzn_results.R
\name{get_mzn_results}
\alias{get_mzn_results}
\title{get the minizinc results from the command line}
\usage{
get_mzn_results(
  model,
  str_options,
  as_string = TRUE,
  as_R6 = FALSE,
  timeout = NA
)
}
\arguments{
\item{model}{the path of the temporary model file.}

\item{str_options}{string returned by solver_options()}

\item{as_string}{bool to specify if output is required as a string.}

\item{as_R6}{bool to specify if output is required as an R6 class object.}

\item{timeout}{time in milis (integer) specifying the timeout limit. NA by default.}
}
\description{
get the minizinc results from the command line
}
