% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfunc_minizinc.R
\name{get_results}
\alias{get_results}
\title{get the results/solution}
\usage{
get_results(
  path_to_model_mzn,
  path_to_data_dzn,
  task = "solve",
  solver = "Gecode",
  path_to_model_fzn = " ",
  path_to_model_ozn = " ",
  FlatZinc_bool = FALSE,
  data_file_bool = TRUE
)
}
\arguments{
\item{path_to_model_mzn}{The path of your model (for eg. model.mzn or model.fzn in case its already compiled) file}

\item{path_to_data_dzn}{The path of data (for eg. data.dzn) file}

\item{task}{The string named task ("compile","standardize" or "solve") that you want to perform.solve by default}

\item{solver}{The name of solver you want to use. Gecode by default}

\item{path_to_model_fzn}{The path to fzn file if model is already compiled. Empty by default}

\item{path_to_model_ozn}{The path to ozn file if there is FlatZinc output. Empty by default}

\item{FlatZinc_bool}{a bool to determine if there is a FlatZinc output file. FALSE by default}

\item{data_file_bool}{a bool to determine if data file was used as input for parameters. TRUE by default}
}
\description{
get the results/solution
}
