% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getMissingPars}
\alias{getMissingPars}
\title{get missing parameters}
\usage{
getMissingPars(
  modelString = "",
  mznpath = "",
  modelStringName = "missing_pars.mzn"
)
}
\arguments{
\item{modelString}{the string representation of a MiniZinc model}

\item{mznpath}{the path of the MiniZinc model mzn file}

\item{modelStringName}{the custom name of the mzn string}
}
\description{
get the names of missing parameters in a model
}
