% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getRModel}
\alias{getRModel}
\title{init all classes}
\usage{
getRModel(mznParseList)
}
\arguments{
\item{mznParseList}{list input}
}
\description{
Given the return value of
`mzn_parse()`, it creates a model in R
using the API mirror
}
