% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_serialize.R
\name{helper_serialize}
\alias{helper_serialize}
\title{helper_serialize (not exposed to the user)}
\usage{
helper_serialize(Exp)
}
\arguments{
\item{Exp}{the expression/value to serialize}
}
\description{
recursive helper function for serializing
constraints and solve items to MiniZinc syntax
}
