% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{helperDeleteExpression}
\alias{helperDeleteExpression}
\title{helper delete expression}
\usage{
helperDeleteExpression(classNm)
}
\arguments{
\item{classNm}{name of the object class}
}
\description{
helper function to search the through a
model for an expression and return the 
object if found
}
