% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{helperDeleteItem}
\alias{helperDeleteItem}
\title{helper delete item}
\usage{
helperDeleteItem(classNm)
}
\arguments{
\item{classNm}{name of the object class}
}
\description{
Helper function to search the through a
model for an item and return the object
if found
}
