% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IncludeItem.R
\name{IncludeItem}
\alias{IncludeItem}
\title{Include Items}
\description{
Include external mzn files
in your model.
}
\section{Super class}{
\code{\link[rminizinc:Item]{rminizinc::Item}} -> \code{IncludeItem}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{.id}}{name of mzn file}

\item{\code{.delete_flag}}{used to delete items}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{.id}}{name of mzn file}

\item{\code{.delete_flag}}{used to delete items}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{IncludeItem$new()}}
\item \href{#method-getmznName}{\code{IncludeItem$getmznName()}}
\item \href{#method-setmznName}{\code{IncludeItem$setmznName()}}
\item \href{#method-c_str}{\code{IncludeItem$c_str()}}
\item \href{#method-getDeleteFlag}{\code{IncludeItem$getDeleteFlag()}}
\item \href{#method-delete}{\code{IncludeItem$delete()}}
\item \href{#method-clone}{\code{IncludeItem$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IncludeItem$new(name = NULL, mzn_str = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{name of the file to include}

\item{\code{mzn_str}}{string representation of Include Item
get file name
set the file name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getmznName"></a>}}
\if{latex}{\out{\hypertarget{method-getmznName}{}}}
\subsection{Method \code{getmznName()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IncludeItem$getmznName()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setmznName"></a>}}
\if{latex}{\out{\hypertarget{method-setmznName}{}}}
\subsection{Method \code{setmznName()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IncludeItem$setmznName(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{name of file}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-c_str"></a>}}
\if{latex}{\out{\hypertarget{method-c_str}{}}}
\subsection{Method \code{c_str()}}{
get the MiniZinc representation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IncludeItem$c_str()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getDeleteFlag"></a>}}
\if{latex}{\out{\hypertarget{method-getDeleteFlag}{}}}
\subsection{Method \code{getDeleteFlag()}}{
delete flag for internal use
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IncludeItem$getDeleteFlag()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete"></a>}}
\if{latex}{\out{\hypertarget{method-delete}{}}}
\subsection{Method \code{delete()}}{
delete the include item
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IncludeItem$delete()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IncludeItem$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
