% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRModel.R
\name{initExpression}
\alias{initExpression}
\title{initExpression (not exposed to the user)}
\usage{
initExpression(pList)
}
\arguments{
\item{pList}{list from mzn_parse to initialise objects}
}
\description{
recursive helper function for initilizing
expression classes
}
