% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{initItem}
\alias{initItem}
\title{initialize R6 from parsed (not to be exposed)}
\usage{
initItem(parsedList)
}
\arguments{
\item{parsedList}{list returned by `mzn_parse()`}
}
\description{
Initialize all the R6 objects
using the list returned by `mzn_parse()` to
create exactly the same structure in R.
}
