% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeclFunctions.R
\name{IntArrDecl}
\alias{IntArrDecl}
\title{new nD int array declaration}
\usage{
IntArrDecl(name, kind, ind, value = NULL, ndim)
}
\arguments{
\item{name}{variable/parameter name}

\item{kind}{"var" or "par"}

\item{ind}{index of the array}

\item{value}{value (NULL by default)}

\item{ndim}{number of dimensions of the array}
}
\description{
declare a new array of int
}
