% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeclFunctions.R
\name{IntDecl}
\alias{IntDecl}
\title{int declaration}
\usage{
IntDecl(name, kind, value = NULL)
}
\arguments{
\item{name}{variable/parameter name}

\item{kind}{"var" or "par"}

\item{value}{value (NULL by default)}
}
\description{
Declare a new int
}
