% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Value.R
\name{IntSetVal}
\alias{IntSetVal}
\title{Integer set value}
\description{
integer range set value in MiniZinc
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{.min}}{minimum value of integer range}

\item{\code{.max}}{maximum value of integer range}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{.min}}{minimum value of integer range}

\item{\code{.max}}{maximum value of integer range}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{IntSetVal$new()}}
\item \href{#method-getMin}{\code{IntSetVal$getMin()}}
\item \href{#method-setMin}{\code{IntSetVal$setMin()}}
\item \href{#method-getMax}{\code{IntSetVal$getMax()}}
\item \href{#method-setMax}{\code{IntSetVal$setMax()}}
\item \href{#method-clone}{\code{IntSetVal$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntSetVal$new(imin, imax)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{imin}}{minimum int value}

\item{\code{imax}}{maximum int value}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getMin"></a>}}
\if{latex}{\out{\hypertarget{method-getMin}{}}}
\subsection{Method \code{getMin()}}{
get the minimum IntVal
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntSetVal$getMin()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setMin"></a>}}
\if{latex}{\out{\hypertarget{method-setMin}{}}}
\subsection{Method \code{setMin()}}{
set the minimum IntVal
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntSetVal$setMin(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{int value to be set}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getMax"></a>}}
\if{latex}{\out{\hypertarget{method-getMax}{}}}
\subsection{Method \code{getMax()}}{
get the maximum IntVal
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntSetVal$getMax()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setMax"></a>}}
\if{latex}{\out{\hypertarget{method-setMax}{}}}
\subsection{Method \code{setMax()}}{
set the maximum IntVal
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntSetVal$setMax(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{int value to be set}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntSetVal$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
