% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Expression.R
\name{Ite}
\alias{Ite}
\title{Ite}
\description{
create if-then-else expressions in MiniZinc
}
\section{Super class}{
\code{\link[rminizinc:Expression]{rminizinc::Expression}} -> \code{Ite}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{.ifs}}{list of if expressions}

\item{\code{.thens}}{list of corresponding then expressions}

\item{\code{.else}}{else expression}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{.ifs}}{list of if expressions}

\item{\code{.thens}}{list of corresponding then expressions}

\item{\code{.else}}{else expression}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Ite$new()}}
\item \href{#method-If}{\code{Ite$If()}}
\item \href{#method-Then}{\code{Ite$Then()}}
\item \href{#method-Else}{\code{Ite$Else()}}
\item \href{#method-c_str}{\code{Ite$c_str()}}
\item \href{#method-clone}{\code{Ite$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ite$new(ifs, thens, Else)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ifs}}{list of if expressions}

\item{\code{thens}}{list of corresponding then expressions}

\item{\code{Else}}{else expression}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-If"></a>}}
\if{latex}{\out{\hypertarget{method-If}{}}}
\subsection{Method \code{If()}}{
get the ith if expression
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ite$If(i)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{i}}{index}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Then"></a>}}
\if{latex}{\out{\hypertarget{method-Then}{}}}
\subsection{Method \code{Then()}}{
get the ith then expression
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ite$Then(i)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{i}}{index}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Else"></a>}}
\if{latex}{\out{\hypertarget{method-Else}{}}}
\subsection{Method \code{Else()}}{
get the else expression
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ite$Else()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-c_str"></a>}}
\if{latex}{\out{\hypertarget{method-c_str}{}}}
\subsection{Method \code{c_str()}}{
get the MiniZinc representation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ite$c_str()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ite$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
