% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iter_delete.R
\name{item_delete}
\alias{item_delete}
\title{search item in model and delete}
\usage{
item_delete(obj)
}
\arguments{
\item{obj}{object to be deleted}
}
\description{
find the object in the model and delete it.
}
