% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{itemDelete}
\alias{itemDelete}
\title{search item in model and delete}
\usage{
itemDelete(classNm, model)
}
\arguments{
\item{classNm}{object to be deleted}

\item{model}{model to delete the object from}
}
\description{
Find the object in the model and delete it.
}
