% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iter_delete.R
\name{iter_expression}
\alias{iter_expression}
\title{iterate through expressions and delete (Under Development)}
\usage{
iter_expression(obj, expObj)
}
\arguments{
\item{obj}{object to delete}

\item{expObj}{expression object to iterate through}
}
\description{
given an object to delete and expression
object, delete all the embedded expression objects that
are identical
}
