% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iter_delete.R
\name{iter_item}
\alias{iter_item}
\title{check all possible items(Under Development)}
\usage{
iter_item(mod, obj, item)
}
\arguments{
\item{mod}{model to be searched}

\item{obj}{object to be deleted}

\item{item}{item to be searched}
}
\description{
find the expressions in the items and delete them if matched
}
