% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{iterExpression}
\alias{iterExpression}
\title{iterate through expressions and delete (Under Development)}
\usage{
iterExpression(classNm, expObj)
}
\arguments{
\item{classNm}{class name of the object to delete}

\item{expObj}{expression object to iterate through}
}
\description{
Given an object to delete and expression object,
delete all the embedded expression objects that
are identical
}
