% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{iterItem}
\alias{iterItem}
\title{check all possible items(Under Development)}
\usage{
iterItem(mod, classNm)
}
\arguments{
\item{mod}{model to be searched}

\item{classNm}{class name of the object to be deleted}
}
\description{
Find the expressions in the items and
delete them if matched
}
