% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommonMZNProblems.R
\name{knapsack}
\alias{knapsack}
\title{knapsack problem}
\usage{
knapsack(n, capacity, profit, size)
}
\arguments{
\item{n}{number of items}

\item{capacity}{total capacity of carrying weight}

\item{profit}{profit corresponding to each item}

\item{size}{weight/size of each item}
}
\description{
Solve a simple knapsack problem (Goal is to maximize the profit)
}
