% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Expression.R
\name{Let}
\alias{Let}
\title{let expression}
\description{
let expression in MiniZinc
}
\section{Super class}{
\code{\link[rminizinc:Expression]{rminizinc::Expression}} -> \code{Let}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{.decl}}{list of local declarations}

\item{\code{.in}}{body of the let}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{.decl}}{list of local declarations}

\item{\code{.in}}{body of the let}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Let$new()}}
\item \href{#method-let}{\code{Let$let()}}
\item \href{#method-body}{\code{Let$body()}}
\item \href{#method-clone}{\code{Let$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Let$new(let, body)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{let}}{list of local declarations}

\item{\code{body}}{body of the let}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-let"></a>}}
\if{latex}{\out{\hypertarget{method-let}{}}}
\subsection{Method \code{let()}}{
access list of local declarations
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Let$let()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-body"></a>}}
\if{latex}{\out{\hypertarget{method-body}{}}}
\subsection{Method \code{body()}}{
return the body
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Let$body()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Let$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
