% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LIBMINIZINC_PATH}
\alias{LIBMINIZINC_PATH}
\title{Flag for internal use}
\format{
A string containing linker flag
}
\usage{
LIBMINIZINC_PATH
}
\description{
Flag for internal use
}
\keyword{datasets}
