% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommonMZNProblems.R
\name{magic_series}
\alias{magic_series}
\title{magic series problem}
\usage{
magic_series(n)
}
\arguments{
\item{n}{order of magic square}
}
\description{
Solve a magic series problem in MiniZinc
Model created by Hakan Kjellerstrand(hakank(at)bonetmail.com)
See : http://www.hakank.org/minizinc/magic_series.mzn
}
