% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommonMZNProblems.R
\name{magic_square}
\alias{magic_square}
\title{magic squares problem}
\usage{
magic_square(n)
}
\arguments{
\item{n}{order of magic square}
}
\description{
Solve a magic squares problem in MiniZinc
Model created by Hakan Kjellerstrand(hakank(at)bonetmail.com)
See : http://www.hakank.org/minizinc/magic_square.mzn
}
