% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{modifyDomain}
\alias{modifyDomain}
\title{update the variable domain}
\usage{
modifyDomain(
  ItemNo,
  NumericSetVal = NULL,
  IdItem = -1L,
  modelString = "",
  mznpath = "",
  modify_mzn = FALSE
)
}
\arguments{
\item{ItemNo}{the item number of the variable whose domain is to be updated.}

\item{NumericSetVal}{list containing the numeric set values. (should be strcitly
of the form c(max = r_value, min = l_value))}

\item{IdItem}{the name of the set variable to be assigned as the set value.}

\item{modelString}{string representation of the MiniZinc model}

\item{mznpath}{path of the mzn file to read the model}

\item{modify_mzn}{if the user wants to modify the mzn parameters.}
}
\description{
update the variable domain
}
