% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{modifyDomainAO}
\alias{modifyDomainAO}
\title{Modify domain using constants}
\usage{
modifyDomainAO(
  ItemNo,
  minVal = NULL,
  maxVal = NULL,
  Val = NULL,
  OPmin = "",
  OPmax = "",
  OP = "",
  modelString = "",
  mznpath = "",
  modify_mzn = FALSE
)
}
\arguments{
\item{ItemNo}{the item number of the variable whose domain is to be updated.}

\item{minVal}{value to add/subtract/multiply/divide to min in domains of type min..max.}

\item{maxVal}{value to add/subtract/multiply/divide to max in domains of type min..max.}

\item{Val}{value to add/subract/multiply/divide to any expression of domain.}

\item{OPmin}{string which specifies the operator to be used with minVal.(PLUS, MINUS,MOD,DIV,MULT)}

\item{OPmax}{string which specifies the operator to be used with maxVal.(PLUS, MINUS,MOD,DIV,MULT)}

\item{OP}{string which specifies the operator to be used with Val.(PLUS, MINUS,MOD,DIV,MULT)}

\item{modelString}{string representation of the MiniZinc model}

\item{mznpath}{path of the mzn file to read the model}

\item{modify_mzn}{if the user wants to modify the mzn parameters.}
}
\description{
Modify domain using constants
}
