% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{modifyDomainFnCall}
\alias{modifyDomainFnCall}
\title{Modify domain function calls}
\usage{
modifyDomainFnCall(
  ItemNo,
  minIdItem = -1L,
  maxIdItem = -1L,
  minFnName = "",
  maxFnName = "",
  modelString = "",
  mznpath = "",
  modify_mzn = FALSE
)
}
\arguments{
\item{ItemNo}{the item number of the variable whose domain is to be updated.}

\item{minIdItem}{min Id item number in domains of type min..max.}

\item{maxIdItem}{max Id item number in domains of type min..max.}

\item{minFnName}{Name of the function to be used for the min ID in min..max.}

\item{maxFnName}{Name of the function to be used for the max ID in min..max.}

\item{modelString}{string representation of the MiniZinc model}

\item{mznpath}{path of the mzn file to read the model}

\item{modify_mzn}{if the user wants to modify the mzn parameters.}
}
\description{
Modify domain function calls
}
