% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{modifyDomainId}
\alias{modifyDomainId}
\title{update the variable domain}
\usage{
modifyDomainId(
  ItemNo,
  maxIdItem = -1L,
  minIdItem = -1L,
  replaceIdItem = -1L,
  modelString = "",
  mznpath = "",
  modify_mzn = FALSE
)
}
\arguments{
\item{ItemNo}{the item number of the variable whose domain is to be updated.}

\item{maxIdItem}{item number of the variable to be assigned as the max domain value.}

\item{minIdItem}{item number of the variable to be assigned as the min domain value.}

\item{replaceIdItem}{item number of the variable to be assigned as the new domain value.}

\item{modelString}{string representation of the MiniZinc model}

\item{mznpath}{path of the mzn file to read the model}

\item{modify_mzn}{if the user wants to modify the mzn parameters.}
}
\description{
update the variable domain
}
