% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{modifyDomainSetVal}
\alias{modifyDomainSetVal}
\title{Modify domain using integer or floating point values}
\usage{
modifyDomainSetVal(
  ItemNo,
  imax = NULL,
  imin = NULL,
  fmin = NULL,
  fmax = NULL,
  modelString = "",
  mznpath = "",
  modify_mzn = FALSE
)
}
\arguments{
\item{ItemNo}{the item number of the variable whose domain is to be updated.}

\item{imax}{maximum integer value of the int set variable domain.}

\item{imin}{minimum integer value of the int set variable domain.}

\item{fmin}{miminum float value of the float set variable domain.}

\item{fmax}{maximum float value of the float set variable domain.}

\item{modelString}{string representation of the MiniZinc model}

\item{mznpath}{path of the mzn file to read the model}

\item{modify_mzn}{if the user wants to modify the mzn parameters.}
}
\description{
Modify domain using integer or floating point values
}
