% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mzn_eval}
\alias{mzn_eval}
\title{MiniZinc model evaluation}
\usage{
mzn_eval(modelString)
}
\arguments{
\item{modelString}{the string representation of the model to be evaluated}
}
\description{
evaluates the MiniZinc model
}
