% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_param.R
\name{new_param}
\alias{new_param}
\title{Declare a new parameter for the MiniZinc model.}
\usage{
new_param(
  model,
  par_name,
  par_type,
  single_par_val,
  enum_val,
  set_val,
  set_type,
  array_val,
  array_type
)
}
\arguments{
\item{model}{an instance of the temporary mzn file}

\item{par_name}{the name of the parameter to initialize}

\item{par_type}{data type of the variable. it can be one of c("int", "float", "bool", "enum", "set", "array" )}

\item{single_par_val}{value of the parameter where par can be "int" , "float" or "bool"}

\item{enum_val}{A vector containing the values of enum}

\item{set_val}{A vector containing the lower and upper limits of a set like c(l, u)}

\item{set_type}{The data type of the set}

\item{array_val}{A vector for 1d array, 2d matrix and higher order matrices for higher dimensions upto 5}

\item{array_type}{The data type of values present in the array}
}
\description{
Declare a new parameter for the MiniZinc model.
}
\details{
The variables declared in.
}
