% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsetoR6.R
\name{parsetoR6}
\alias{parsetoR6}
\title{parsetoR6 class}
\description{
class for parsing MiniZinc output strings and storing them as R6 objects.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{optimal_solution}}{stores the optimal solution. NULL if solution not found}

\item{\code{best_solution}}{stores the optimal solution. NULL if solution not found}

\item{\code{solutions}}{stores all the possible solutions.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{parsetoR6$new()}}
\item \href{#method-clone}{\code{parsetoR6$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
initialize the class objects
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{parsetoR6$new(mzn_result)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mzn_result}}{the result (string output) of MiniZinc model}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{parsetoR6$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
