% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommonMZNProblems.R
\name{production_planning}
\alias{production_planning}
\title{production planning problem}
\usage{
production_planning(
  nproducts,
  profit,
  pnames,
  nresources,
  capacity,
  rnames,
  consumption
)
}
\arguments{
\item{nproducts}{number of different products}

\item{profit}{profit for each product (1-D vector)}

\item{pnames}{names of each product (1-D vector)}

\item{nresources}{number of resources}

\item{capacity}{amount of each resource available (1-D vector)}

\item{rnames}{names of each resource (1-D vector)}

\item{consumption}{units of each resource required to produce 
1 unit of product (2-D vector to be provided as 1-D vector)}
}
\description{
simple production planning problem taken from
https://github.com/MiniZinc/minizinc-examples
Goal is to maximize the profit
}
