% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PROJECT_DIRECTORY}
\alias{PROJECT_DIRECTORY}
\title{Absolute path of project directory}
\format{
A string containing absolute path of the project directory
}
\usage{
PROJECT_DIRECTORY
}
\description{
Absolute path of project directory
}
\keyword{datasets}
