% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{rminizinc-package}
\alias{rminizinc}
\alias{rminizinc-package}
\title{rminizinc: R Interface to Use 'MiniZinc'}
\description{
Load the required libraries used by 
most of the functions and classes
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/acharaakshit/RMiniZinc}
  \item Report bugs at \url{https://github.com/acharaakshit/RMiniZinc/issues}
}

}
