% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable.R
\name{set_checks}
\alias{set_checks}
\title{set checks}
\description{
this class contains the various checks required for sets
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{set_checks$new()}}
\item \href{#method-clone}{\code{set_checks$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
checks related to set variables and parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{set_checks$new(set_kind, set_type, set_value, set_expression)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{set_kind}}{parameter or decision}

\item{\code{set_type}}{data type of the set}

\item{\code{set_value}}{the value of set if it's a parameter}

\item{\code{set_expression}}{the input expression}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{set_checks$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
