% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfunc_minizinc.R
\name{setup_checker_model_input}
\alias{setup_checker_model_input}
\title{A function to setup the checker file}
\usage{
setup_checker_model_input(dec_var_names, dec_var_types, constraint_list, path)
}
\arguments{
\item{constraint_list}{list of constraints of the model.}

\item{path}{th epath of model_mzc_mzn file}

\item{dec_vars}{the decision variables to be written to the checker file.}
}
\description{
A function to setup the checker file
}
