% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfunc_minizinc.R
\name{setup_constraints}
\alias{setup_constraints}
\title{setup the constraints of the model. structure of output should look like constraint 〈Boolean expression〉;}
\usage{
setup_constraints(constraint_list, append_bool = TRUE, path)
}
\arguments{
\item{constraint_list}{list of the constraints of the model}

\item{append_bool}{bool to tell if it is required to open the file in append mode. TRUE by default}

\item{path}{the oath of the model_mzn file}
}
\description{
setup the constraints of the model. structure of output should look like constraint 〈Boolean expression〉;
}
