% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfunc_minizinc.R
\name{setup_output}
\alias{setup_output}
\title{setup the output scheme}
\usage{
setup_output(
  output_vars = c(),
  append_bool = TRUE,
  default_output_bool = TRUE,
  path
)
}
\arguments{
\item{output_vars}{the list of output variables, i.e the variables you want to print}

\item{append_bool}{bool to tell if it is required to open the file in append mode. TRUE by default}

\item{default_output_bool}{bool to determine if you want to output all the decision variables.TRUE by default.}

\item{path}{the path of model mzn}
}
\description{
setup the output scheme
}
