% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfunc_minizinc.R
\name{setup_solvers}
\alias{setup_solvers}
\title{mention the type of problem}
\usage{
setup_solvers(
  type_of_problem,
  arithmetic_expression = " ",
  append_bool = TRUE,
  path
)
}
\arguments{
\item{type_of_problem}{the type of optimization problem. choose s for satisfaction problem,
min for minimization problem and max for maximization
@param arithmetic_expression string containing arithmetic expression for maximization and minimization
problems. Empty by default.}

\item{append_bool}{bool to tell if it is required to open the file in append mode. TRUE by default}

\item{path}{the path of model mzn}
}
\description{
mention the type of problem
}
