% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Value.R
\name{SetVal}
\alias{SetVal}
\title{SetVal class}
\description{
possible types of set values in MiniZinc
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{.isv}}{the int set value}

\item{\code{.fsv}}{the float set value}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{.isv}}{the int set value}

\item{\code{.fsv}}{the float set value}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SetVal$new()}}
\item \href{#method-isv}{\code{SetVal$isv()}}
\item \href{#method-fsv}{\code{SetVal$fsv()}}
\item \href{#method-clone}{\code{SetVal$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SetVal$new(val)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-isv"></a>}}
\if{latex}{\out{\hypertarget{method-isv}{}}}
\subsection{Method \code{isv()}}{
return the int set value
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SetVal$isv()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fsv"></a>}}
\if{latex}{\out{\hypertarget{method-fsv}{}}}
\subsection{Method \code{fsv()}}{
return the float set value
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SetVal$fsv()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SetVal$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
