% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sol_parse}
\alias{sol_parse}
\title{parse the solution}
\usage{
sol_parse(solutionString)
}
\arguments{
\item{solutionString}{solution of the model as a string representation}
}
\description{
can parse the solution of a model
}
