% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SolveType.R
\name{SolveItem}
\alias{SolveItem}
\title{mention if the problem is satisfaction, minimization or maximization.}
\description{
This is a class that will be used to specify whether the optimization problem is a satisfaction,
minimization or maximization problem
}
\section{Super class}{
\code{\link[rminizinc:Item]{rminizinc::Item}} -> \code{SolveType}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SolveItem$new()}}
\item \href{#method-e}{\code{SolveItem$e()}}
\item \href{#method-st}{\code{SolveItem$st()}}
\item \href{#method-c_str}{\code{SolveItem$c_str()}}
\item \href{#method-clone}{\code{SolveItem$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
create an instance of specify_problem class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolveItem$new(solve_type, expression = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{solve_type}}{satisfaction, minimization or maximization}

\item{\code{expression}}{expression to minimize or maximize}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-e"></a>}}
\if{latex}{\out{\hypertarget{method-e}{}}}
\subsection{Method \code{e()}}{
return the expression
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolveItem$e()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-st"></a>}}
\if{latex}{\out{\hypertarget{method-st}{}}}
\subsection{Method \code{st()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolveItem$st()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-c_str"></a>}}
\if{latex}{\out{\hypertarget{method-c_str}{}}}
\subsection{Method \code{c_str()}}{
to string method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolveItem$c_str()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolveItem$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
