% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SOLVER_BIN}
\alias{SOLVER_BIN}
\title{Path of the solver executable directory}
\format{
A string containing path of solver executable directory
}
\usage{
SOLVER_BIN
}
\description{
Path of the solver executable directory
}
\keyword{datasets}
