% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCollections.R
\name{stringExpressions}
\alias{stringExpressions}
\title{get strings}
\usage{
stringExpressions(vals)
}
\arguments{
\item{vals}{vector of string values}
}
\description{
get a list of string expressions
}
