% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeclFunctions.R
\name{StringSetDecl}
\alias{StringSetDecl}
\title{set of string declaration}
\usage{
StringSetDecl(name, kind, value = NULL)
}
\arguments{
\item{name}{variable/parameter name}

\item{kind}{"var" or "par"}

\item{value}{value of the set (or NULL)}
}
\description{
declare a new set of string
}
