% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type.R
\name{Type}
\alias{Type}
\title{Type class}
\description{
the type information of the class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{.bt}}{the base type}

\item{\code{.kind}}{parameter or decision}

\item{\code{.dim}}{the number of dimensions}

\item{\code{.typeInst}}{the type instantiation}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{.bt}}{the base type}

\item{\code{.kind}}{parameter or decision}

\item{\code{.dim}}{the number of dimensions}

\item{\code{.typeInst}}{the type instantiation}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Type$new()}}
\item \href{#method-bt}{\code{Type$bt()}}
\item \href{#method-kind}{\code{Type$kind()}}
\item \href{#method-ti}{\code{Type$ti()}}
\item \href{#method-clone}{\code{Type$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Type$new(base_type, kind, type_inst = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{base_type}}{the base type}

\item{\code{kind}}{parameter or decision}

\item{\code{type_inst}}{indices or domain}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-bt"></a>}}
\if{latex}{\out{\hypertarget{method-bt}{}}}
\subsection{Method \code{bt()}}{
return the base type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Type$bt()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-kind"></a>}}
\if{latex}{\out{\hypertarget{method-kind}{}}}
\subsection{Method \code{kind()}}{
return the kind
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Type$kind()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ti"></a>}}
\if{latex}{\out{\hypertarget{method-ti}{}}}
\subsection{Method \code{ti()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Type$ti()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Type$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
