% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mzn.R
\name{type_of}
\alias{type_of}
\title{helper function to get data type}
\usage{
type_of(val)
}
\arguments{
\item{val}{the value of parameter}
}
\description{
This function is used to map the data types obtained using typeof()
in R to the data types of MiniZinc
}
