% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable.R
\name{VarDecl}
\alias{VarDecl}
\title{class for variable declaration}
\description{
Converts a data type into a variable declaration
}
\section{Super class}{
\code{\link[rminizinc:Expression]{rminizinc::Expression}} -> \code{VarDecl}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{.type}}{type information}

\item{\code{id}}{name of the variable}

\item{\code{.expression}}{the initialization expression}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{.type}}{type information}

\item{\code{id}}{name of the variable}

\item{\code{.expression}}{the initialization expression}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{VarDecl$new()}}
\item \href{#method-get_id}{\code{VarDecl$get_id()}}
\item \href{#method-isPar}{\code{VarDecl$isPar()}}
\item \href{#method-isVar}{\code{VarDecl$isVar()}}
\item \href{#method-e}{\code{VarDecl$e()}}
\item \href{#method-clone}{\code{VarDecl$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
initialize the VarDecl constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VarDecl$new(type, expression = NULL, id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{the type object}

\item{\code{expression}}{the object of a data type in MiniZinc. NULL by default}

\item{\code{id}}{the id/name of the declared variable}

\item{\code{kind}}{parameter or decision}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_id"></a>}}
\if{latex}{\out{\hypertarget{method-get_id}{}}}
\subsection{Method \code{get_id()}}{

\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VarDecl$get_id()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-isPar"></a>}}
\if{latex}{\out{\hypertarget{method-isPar}{}}}
\subsection{Method \code{isPar()}}{
check if it's a parameter
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VarDecl$isPar()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-isVar"></a>}}
\if{latex}{\out{\hypertarget{method-isVar}{}}}
\subsection{Method \code{isVar()}}{
check if it's a decision variable
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VarDecl$isVar()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-e"></a>}}
\if{latex}{\out{\hypertarget{method-e}{}}}
\subsection{Method \code{e()}}{
return the initialization expression
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VarDecl$e()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VarDecl$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
