% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeclFunctions.R
\name{VarDomainDecl}
\alias{VarDomainDecl}
\title{declare 0-D variable with domain}
\usage{
VarDomainDecl(name, dom)
}
\arguments{
\item{name}{variable name}

\item{dom}{domain}
}
\description{
Declare a 0 dimensional (int, float,
bool or string) variable with domain
}
