% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mzn.R
\name{write_mzn}
\alias{write_mzn}
\title{write the model to mzn file}
\usage{
write_mzn(model, mzn)
}
\arguments{
\item{model}{an object of the model class.}

\item{mzn}{the path of temporary model.}
}
\description{
write all the components of the minizinc model to the 
temporary mzn file
}
