% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Expression.R
\name{Array}
\alias{Array}
\title{create an array}
\description{
create an array in MiniZinc
}
\section{Super class}{
\code{\link[rminizinc:Expression]{rminizinc::Expression}} -> \code{Array}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{.exprVec}}{vector of value expressions}

\item{\code{.dims}}{vector of dimension expressions}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{.exprVec}}{vector of value expressions}

\item{\code{.dims}}{vector of dimension expressions}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Array$new()}}
\item \href{#method-ndims}{\code{Array$ndims()}}
\item \href{#method-getDim}{\code{Array$getDim()}}
\item \href{#method-c_str}{\code{Array$c_str()}}
\item \href{#method-clone}{\code{Array$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
constructor for an int literal
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Array$new(exprVec, dims)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{exprVec}}{the value of the integer}

\item{\code{dims}}{dimension expression vector}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ndims"></a>}}
\if{latex}{\out{\hypertarget{method-ndims}{}}}
\subsection{Method \code{ndims()}}{
get the number of dimensions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Array$ndims()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getDim"></a>}}
\if{latex}{\out{\hypertarget{method-getDim}{}}}
\subsection{Method \code{getDim()}}{
get the ith dimension vector
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Array$getDim(i)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{i}}{index}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-c_str"></a>}}
\if{latex}{\out{\hypertarget{method-c_str}{}}}
\subsection{Method \code{c_str()}}{
return the MiniZinc representation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Array$c_str()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Array$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
