% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeclFunctions.R
\name{ArrDomainDecl}
\alias{ArrDomainDecl}
\title{declare n-D array with domain}
\usage{
ArrDomainDecl(name, kind, dom, ndim)
}
\arguments{
\item{name}{variable name}

\item{kind}{variable or parameter}

\item{dom}{domain}

\item{ndim}{number of dimensions}
}
\description{
Declare a n-dimensional array 
with domain
}
