% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Expression.R
\name{Comprehension}
\alias{Comprehension}
\title{Comprehension class}
\description{
create a Comprehension
}
\section{Super class}{
\code{\link[rminizinc:Expression]{rminizinc::Expression}} -> \code{Comprehension}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{.generators}}{a vector of generators}

\item{\code{.expression}}{the comprehension expression}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{.generators}}{a vector of generators}

\item{\code{.expression}}{the comprehension expression}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Comprehension$new()}}
\item \href{#method-ngens}{\code{Comprehension$ngens()}}
\item \href{#method-gen_i}{\code{Comprehension$gen_i()}}
\item \href{#method-In}{\code{Comprehension$In()}}
\item \href{#method-where}{\code{Comprehension$where()}}
\item \href{#method-decl}{\code{Comprehension$decl()}}
\item \href{#method-e}{\code{Comprehension$e()}}
\item \href{#method-c_str}{\code{Comprehension$c_str()}}
\item \href{#method-clone}{\code{Comprehension$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Comprehension$new(generators, e)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{generators}}{generators of the expression}

\item{\code{e}}{inside the comprehension}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ngens"></a>}}
\if{latex}{\out{\hypertarget{method-ngens}{}}}
\subsection{Method \code{ngens()}}{
get the number of generators
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Comprehension$ngens()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gen_i"></a>}}
\if{latex}{\out{\hypertarget{method-gen_i}{}}}
\subsection{Method \code{gen_i()}}{
get the ith generator expression
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Comprehension$gen_i(i)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{i}}{index}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-In"></a>}}
\if{latex}{\out{\hypertarget{method-In}{}}}
\subsection{Method \code{In()}}{
get the in expression of ith generator
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Comprehension$In(i)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{i}}{index}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-where"></a>}}
\if{latex}{\out{\hypertarget{method-where}{}}}
\subsection{Method \code{where()}}{
get the where expression of ith generator
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Comprehension$where(i)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{i}}{index}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-decl"></a>}}
\if{latex}{\out{\hypertarget{method-decl}{}}}
\subsection{Method \code{decl()}}{
get the idecl declaration of igen generator
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Comprehension$decl(igen, idecl)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{igen}}{generator index}

\item{\code{idecl}}{declaration index}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-e"></a>}}
\if{latex}{\out{\hypertarget{method-e}{}}}
\subsection{Method \code{e()}}{
get the expression
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Comprehension$e()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-c_str"></a>}}
\if{latex}{\out{\hypertarget{method-c_str}{}}}
\subsection{Method \code{c_str()}}{
get the MiniZinc representation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Comprehension$c_str()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Comprehension$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
