% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_decision_var.R
\name{create_decision_vars}
\alias{create_decision_vars}
\title{create the variables.}
\usage{
create_decision_vars(
  single_dec_var_names,
  single_dec_var_ranges,
  dec_var_collection_names = c(),
  dec_collection_types = c(),
  dec_var_collection_types = c(),
  dec_var_array_ranges = c(),
  model
)
}
\arguments{
\item{single_dec_var_names}{list of names of single decision variables}

\item{single_dec_var_ranges}{list of ranges of the single decision variables. Typical range should look like c(2,10).}

\item{dec_var_collection_names}{list of names of collection decision variables(array or set)}

\item{dec_collection_types}{list of types of collection.Possible types are "array" or "set".}

\item{dec_var_collection_types}{list of data types of collection of decision variables.}

\item{dec_var_array_ranges}{list of array ranges in order of  the array variables used.typical element in the list
should look like c(lower_limit,upper_limit), for eg. c(2,10) for 2D array and single numbers for eg 7 for
1D array .}

\item{model}{the path for model mzn}
}
\description{
create the variables.
}
