% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeclFunctions.R
\name{FloatDecl}
\alias{FloatDecl}
\title{float declaration}
\usage{
FloatDecl(name, kind, value = NULL, domain = NULL)
}
\arguments{
\item{name}{variable/parameter name}

\item{kind}{"var" or "par"}

\item{value}{value (NULL by default)}

\item{domain}{domain of the float variable}
}
\description{
Declare a new float
}
