% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{FunctionItem}
\alias{FunctionItem}
\title{Function Items}
\description{
independent functions (that are not part of any other items)
in a MiniZinc model
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{.id}}{name of the function}

\item{\code{.e}}{expression in the function}

\item{\code{.params}}{parameter declarations}

\item{\code{.ann}}{annotation}

\item{\code{.rt}}{return type of the function}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{.id}}{name of the function}

\item{\code{.e}}{expression in the function}

\item{\code{.params}}{parameter declarations}

\item{\code{.ann}}{annotation}

\item{\code{.rt}}{return type of the function}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FunctionItem$new()}}
\item \href{#method-name}{\code{FunctionItem$name()}}
\item \href{#method-b}{\code{FunctionItem$b()}}
\item \href{#method-ann}{\code{FunctionItem$ann()}}
\item \href{#method-rtype}{\code{FunctionItem$rtype()}}
\item \href{#method-c_str}{\code{FunctionItem$c_str()}}
\item \href{#method-clone}{\code{FunctionItem$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FunctionItem$new(name, pars, rt, ann = NULL, body = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{name of the function}

\item{\code{pars}}{parameter declarations}

\item{\code{rt}}{the return type ("bool par", "bool var" or other)}

\item{\code{ann}}{annotation}

\item{\code{body}}{body of the function}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-name"></a>}}
\if{latex}{\out{\hypertarget{method-name}{}}}
\subsection{Method \code{name()}}{
get the name of the function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FunctionItem$name()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-b"></a>}}
\if{latex}{\out{\hypertarget{method-b}{}}}
\subsection{Method \code{b()}}{
get the function body
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FunctionItem$b()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ann"></a>}}
\if{latex}{\out{\hypertarget{method-ann}{}}}
\subsection{Method \code{ann()}}{
get the function annotation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FunctionItem$ann()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rtype"></a>}}
\if{latex}{\out{\hypertarget{method-rtype}{}}}
\subsection{Method \code{rtype()}}{
get if the function is a test, predicate 
or a function call itself.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FunctionItem$rtype()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-c_str"></a>}}
\if{latex}{\out{\hypertarget{method-c_str}{}}}
\subsection{Method \code{c_str()}}{
get the MiniZinc representation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FunctionItem$c_str()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FunctionItem$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
