% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_missing_pars.R
\name{get_missing_pars}
\alias{get_missing_pars}
\title{get missing parameters}
\usage{
get_missing_pars(model)
}
\arguments{
\item{model}{object of Model class}
}
\description{
Get the values of the missing parameters
}
