% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getType}
\alias{getType}
\title{initialized type (not exposed to user)}
\usage{
getType(type_str, kind)
}
\arguments{
\item{type_str}{type string returned by `parse_mzn()`.}

\item{kind}{par or var}
}
\description{
Helper function to initialise the type.
}
