% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeclFunctions.R
\name{IntArrDecl}
\alias{IntArrDecl}
\title{new 1D int array declaration}
\usage{
IntArrDecl(name, kind, ind, e = NULL)
}
\arguments{
\item{name}{variable/parameter name}

\item{kind}{"var" or "par"}

\item{ind}{index of the array}

\item{e}{value (NULL by default)}
}
\description{
declare a new array of int
}
