% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LIBMINIZINC_PATH}
\alias{LIBMINIZINC_PATH}
\title{Absolute path of the libminizinc library}
\format{
A string containing linker flag
}
\usage{
LIBMINIZINC_PATH
}
\description{
Absolute path of the libminizinc library
}
\keyword{datasets}
