% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mzn_eval}
\alias{mzn_eval}
\title{MiniZinc model evaluation}
\usage{
mzn_eval(
  solver,
  lib_path,
  r_model,
  dzn_path = "",
  all_solutions = TRUE,
  time_limit = 300000L
)
}
\arguments{
\item{solver}{the name of the solver to use.}

\item{lib_path}{the path of the library where the solver is present.}

\item{r_model}{R6 Model object}

\item{dzn_path}{path of the datafile to be used.}

\item{all_solutions}{bool to specify if all solutions are specified.}

\item{time_limit}{stop after <time_limit> milliseconds}
}
\description{
evaluates the MiniZinc model
}
