% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mzn_eval}
\alias{mzn_eval}
\title{MiniZinc model evaluation}
\usage{
mzn_eval(
  solver,
  libpath,
  modelString = "",
  mznpath = "",
  dznpath = "",
  all_solutions = TRUE
)
}
\arguments{
\item{solver}{the name of the solver to use.}

\item{libpath}{the path of the library where the solver is present.}

\item{modelString}{the string representation of the model to be evaluated.}

\item{mznpath}{the path of the MiniZinc model file.}

\item{dznpath}{path of the datafile to be used.}

\item{all_solutions}{bool to specify if all solutions are specified.}
}
\description{
evaluates the MiniZinc model
}
