% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mzn_eval}
\alias{mzn_eval}
\title{MiniZinc model evaluation}
\usage{
mzn_eval(
  lib_path = "",
  r_model = NULL,
  mzn_path = "",
  model_string = "",
  solver = "org.gecode.gecode",
  dzn_path = "",
  all_solutions = TRUE,
  time_limit = 300000L
)
}
\arguments{
\item{lib_path}{the path of the library where the solver is present.}

\item{r_model}{R6 Model object}

\item{mzn_path}{path of the mzn file to be solved}

\item{model_string}{model string to be solved.}

\item{solver}{the name of the solver to use.}

\item{dzn_path}{path of the datafile to be used.}

\item{all_solutions}{bool to specify if all solutions are specified.}

\item{time_limit}{stop after <time_limit> milliseconds}
}
\description{
evaluates the MiniZinc model
}
