% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mzn_parse}
\alias{mzn_parse}
\title{MiniZinc syntax parser}
\usage{
mzn_parse(modelString)
}
\arguments{
\item{modelString}{string representation of the MiniZinc model}
}
\description{
parses the MiniZinc syntax into R objects
}
