% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mzn_parse}
\alias{mzn_parse}
\title{MiniZinc syntax parser}
\usage{
mzn_parse(
  modelString = "",
  mznpath = "",
  modelStringName = "mzn_parse.mzn",
  includePath = NULL
)
}
\arguments{
\item{modelString}{string representation of the MiniZinc model.}

\item{mznpath}{the path of model mzn.}

\item{modelStringName}{the name of model string.}

\item{includePath}{path of the included mzn in the model if it exists.}
}
\description{
parses the MiniZinc syntax into R objects
}
