% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mzn_parse}
\alias{mzn_parse}
\title{MiniZinc syntax parser}
\usage{
mzn_parse(
  modData,
  modelString = "",
  mznfilename = as.character(c()),
  dznfilename = as.character(c()),
  modelStringName = "abc.mzn"
)
}
\arguments{
\item{modData}{list containing the parameter values.}

\item{modelString}{string representation of the MiniZinc model.}

\item{mznfilename}{the name of model.}

\item{dznfilename}{the name of the dzn file.}

\item{modelStringName}{the name of model string.}
}
\description{
parses the MiniZinc syntax into R objects
}
