% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mzn_parse}
\alias{mzn_parse}
\title{MiniZinc syntax parser}
\usage{
mzn_parse(model_string = "", mzn_path = "", include_path = NULL)
}
\arguments{
\item{model_string}{string representation of the MiniZinc model.}

\item{mzn_path}{the path of model mzn.}

\item{include_path}{path of the included mzn in the model if it exists.}
}
\description{
parses the MiniZinc syntax into R objects
}
