% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{results}
\alias{results}
\title{get the minizinc results.}
\description{
This class will get the MiniZinc results by creating the mzn file.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{mzn}}{temporary file for writing a model}

\item{\code{model}}{MiniZinc \code{\link{model}}}

\item{\code{result_type}}{specify whether output is required as string or R6}

\item{\code{result}}{the result of the model}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{results$new()}}
\item \href{#method-clone}{\code{results$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
obtain the results from the model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{results$new(
  model,
  result_type,
  all_solutions = FALSE,
  statistics = FALSE,
  threads = 1,
  keep_files = FALSE,
  timeout = NA
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{object of the model class}

\item{\code{result_type}}{string or R6 MiniZinc output}

\item{\code{all_solutions}}{boolean to specify is all solutions are needed. FALSE by default.}

\item{\code{statistics}}{boolean to specify if statistics are required. FALSE by default.}

\item{\code{threads}}{number of threads to be used by MiniZinc. 1 by default.}

\item{\code{keep_files}}{bool to specify if the flattened model should be kept or not. FALSE by default.}

\item{\code{timeout}}{time in milis (integer) specifying the timeout limit. NA by default.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{results$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
