% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{set_params}
\alias{set_params}
\title{assign the missing parameters to the model}
\usage{
set_params(modData, modelString)
}
\arguments{
\item{modData}{list containing the parameter values.}

\item{modelString}{string representation of the MiniZinc model}
}
\description{
assign the missing parameters to the model
}
