% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{set_params}
\alias{set_params}
\title{assign values to the missing parameters}
\usage{
set_params(
  modData,
  modelString = "",
  mznpath = "",
  modify_mzn = FALSE,
  includePath = NULL
)
}
\arguments{
\item{modData}{list containing the parameter values.}

\item{modelString}{string representation of the MiniZinc model}

\item{mznpath}{path of the mzn file to read the model}

\item{modify_mzn}{if the user wants to modify the mzn parameters.}

\item{includePath}{path of the included mzn in the model if it exists.}
}
\description{
assign values to the missing parameters
}
