% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_params.R
\name{set_params}
\alias{set_params}
\title{set missing parameters}
\usage{
set_params(model, modData)
}
\arguments{
\item{model}{Model object}

\item{modData}{list of the value objects to be assigned}
}
\description{
Assign values to parameters which don't have
a value assigned yet.
}
