% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_constraints.R
\name{setup_constraints}
\alias{setup_constraints}
\title{specify the constraints of the model.}
\usage{
setup_constraints(model, constraint_list)
}
\arguments{
\item{model}{the oath of the model_mzn file}

\item{constraint_list}{list of the constraints of the model}
}
\description{
specify the constraints of the model.
}
