% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SolveType.R
\name{SolveItem}
\alias{SolveItem}
\title{mention if the problem is satisfaction, minimization or maximization.}
\description{
This is a class that will be used to specify whether the optimization problem is a satisfaction,
minimization or maximization problem
}
\section{Super class}{
\code{\link[rminizinc:Item]{rminizinc::Item}} -> \code{SolveType}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{.ann}}{annotation of the solve type}

\item{\code{.delete_flag}}{used to delete items}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{.ann}}{annotation of the solve type}

\item{\code{.delete_flag}}{used to delete items}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SolveItem$new()}}
\item \href{#method-e}{\code{SolveItem$e()}}
\item \href{#method-st}{\code{SolveItem$st()}}
\item \href{#method-c_str}{\code{SolveItem$c_str()}}
\item \href{#method-getDeleteFlag}{\code{SolveItem$getDeleteFlag()}}
\item \href{#method-delete}{\code{SolveItem$delete()}}
\item \href{#method-clone}{\code{SolveItem$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
create an instance of specify_problem class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolveItem$new(solve_type, e = NULL, ann = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{solve_type}}{satisfaction, minimization or maximization}

\item{\code{e}}{expression to minimize or maximize}

\item{\code{ann}}{annotation}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-e"></a>}}
\if{latex}{\out{\hypertarget{method-e}{}}}
\subsection{Method \code{e()}}{
return the expression
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolveItem$e()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-st"></a>}}
\if{latex}{\out{\hypertarget{method-st}{}}}
\subsection{Method \code{st()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolveItem$st()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-c_str"></a>}}
\if{latex}{\out{\hypertarget{method-c_str}{}}}
\subsection{Method \code{c_str()}}{
to string method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolveItem$c_str()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getDeleteFlag"></a>}}
\if{latex}{\out{\hypertarget{method-getDeleteFlag}{}}}
\subsection{Method \code{getDeleteFlag()}}{
delete flag for internal use
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolveItem$getDeleteFlag()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete"></a>}}
\if{latex}{\out{\hypertarget{method-delete}{}}}
\subsection{Method \code{delete()}}{
delete the solve item
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolveItem$delete()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolveItem$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
