% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver_options.R
\name{solver_options}
\alias{solver_options}
\title{Specify the solver options}
\usage{
solver_options(
  solver = "Gecode",
  all_solutions = FALSE,
  statistics = FALSE,
  keep_files = FALSE,
  threads = 1
)
}
\arguments{
\item{solver}{name of the solver to be used by MiniZinc "Gecode" by default.}

\item{all_solutions}{bool to specify if all the solutions should be printed or not. FALSE by default.}

\item{statistics}{bool to specify if the performance statistics should be printed or not. FALSE by default.}

\item{keep_files}{bool to specify if the flattened model should be kept or not. FALSE by default.}

\item{threads}{integer specifying the number of threads to use. 1 by default.}
}
\description{
Specify the solver options
}
