% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_problem.R
\name{specify_problem}
\alias{specify_problem}
\title{mention the type of problem}
\usage{
specify_problem(model, type_of_problem, arithmetic_expression = " ")
}
\arguments{
\item{model}{the path of model mzn}

\item{type_of_problem}{the type of optimization problem. choose s for satisfaction problem,
min for minimization problem and max for maximization}

\item{arithmetic_expression}{string containing arithmetic expression for maximization and minimization
problems. Empty by default.}
}
\description{
mention the type of problem
}
