% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Type.R
\name{Type}
\alias{Type}
\title{Type class}
\description{
The information of different data types
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{.bt}}{the base type}

\item{\code{.kind}}{parameter or decision}

\item{\code{.dim}}{the number of dimensions
set or plain}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{.bt}}{the base type}

\item{\code{.kind}}{parameter or decision}

\item{\code{.dim}}{the number of dimensions
set or plain}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Type$new()}}
\item \href{#method-bt}{\code{Type$bt()}}
\item \href{#method-st}{\code{Type$st()}}
\item \href{#method-kind}{\code{Type$kind()}}
\item \href{#method-ndim}{\code{Type$ndim()}}
\item \href{#method-isInt}{\code{Type$isInt()}}
\item \href{#method-isFloat}{\code{Type$isFloat()}}
\item \href{#method-isBool}{\code{Type$isBool()}}
\item \href{#method-isString}{\code{Type$isString()}}
\item \href{#method-isSet}{\code{Type$isSet()}}
\item \href{#method-isIntSet}{\code{Type$isIntSet()}}
\item \href{#method-isFloatSet}{\code{Type$isFloatSet()}}
\item \href{#method-isBoolSet}{\code{Type$isBoolSet()}}
\item \href{#method-clone}{\code{Type$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Type$new(base_type, kind, dim = 0, set_type = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{base_type}}{the base type}

\item{\code{kind}}{parameter or decision}

\item{\code{dim}}{the number of dimensions}

\item{\code{set_type}}{set or plain}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-bt"></a>}}
\if{latex}{\out{\hypertarget{method-bt}{}}}
\subsection{Method \code{bt()}}{
return the base type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Type$bt()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-st"></a>}}
\if{latex}{\out{\hypertarget{method-st}{}}}
\subsection{Method \code{st()}}{
return if it's set type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Type$st()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-kind"></a>}}
\if{latex}{\out{\hypertarget{method-kind}{}}}
\subsection{Method \code{kind()}}{
return the kind
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Type$kind()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ndim"></a>}}
\if{latex}{\out{\hypertarget{method-ndim}{}}}
\subsection{Method \code{ndim()}}{
return the number of dimensions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Type$ndim()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-isInt"></a>}}
\if{latex}{\out{\hypertarget{method-isInt}{}}}
\subsection{Method \code{isInt()}}{
check if it's an int
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Type$isInt()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-isFloat"></a>}}
\if{latex}{\out{\hypertarget{method-isFloat}{}}}
\subsection{Method \code{isFloat()}}{
check if it's a float
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Type$isFloat()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-isBool"></a>}}
\if{latex}{\out{\hypertarget{method-isBool}{}}}
\subsection{Method \code{isBool()}}{
check if it's a bool
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Type$isBool()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-isString"></a>}}
\if{latex}{\out{\hypertarget{method-isString}{}}}
\subsection{Method \code{isString()}}{
check if it's a string
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Type$isString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-isSet"></a>}}
\if{latex}{\out{\hypertarget{method-isSet}{}}}
\subsection{Method \code{isSet()}}{
return if set in MiniZinc
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Type$isSet()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-isIntSet"></a>}}
\if{latex}{\out{\hypertarget{method-isIntSet}{}}}
\subsection{Method \code{isIntSet()}}{
check if it's a set of int
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Type$isIntSet()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-isFloatSet"></a>}}
\if{latex}{\out{\hypertarget{method-isFloatSet}{}}}
\subsection{Method \code{isFloatSet()}}{
check if it's a set of float
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Type$isFloatSet()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-isBoolSet"></a>}}
\if{latex}{\out{\hypertarget{method-isBoolSet}{}}}
\subsection{Method \code{isBoolSet()}}{
check if it's a set of bool
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Type$isBoolSet()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Type$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
