% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Expression.R
\name{TypeInst}
\alias{TypeInst}
\title{TypeInst class}
\description{
type instantiation with indices, etc.
}
\section{Super class}{
\code{\link[rminizinc:Expression]{rminizinc::Expression}} -> \code{TypeInst}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{.indExpr}}{the index expression}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{.indExpr}}{the index expression}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TypeInst$new()}}
\item \href{#method-ranges}{\code{TypeInst$ranges()}}
\item \href{#method-clone}{\code{TypeInst$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
constuctor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TypeInst$new(indexExprVec)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{indexExprVec}}{the expression vector of indices}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ranges"></a>}}
\if{latex}{\out{\hypertarget{method-ranges}{}}}
\subsection{Method \code{ranges()}}{
return the index expression vector
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TypeInst$ranges()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TypeInst$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
