% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Expression.R
\name{TypeInst}
\alias{TypeInst}
\title{TypeInst}
\description{
Create type instantiation with indices, etc.
}
\examples{
TypeInst$new(type = Type$new(base_type = "int", kind = "par" ,dim = 1), 
             domain = Set$new(IntSetVal$new(2,5)))
}
\section{Super class}{
\code{\link[rminizinc:Expression]{rminizinc::Expression}} -> \code{TypeInst}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{.indExpr}}{the index expression}

\item{\code{.domain}}{the domain of possible values to be taken}

\item{\code{.type}}{the type information}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{.indExpr}}{the index expression}

\item{\code{.domain}}{the domain of possible values to be taken}

\item{\code{.type}}{the type information}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TypeInst$new()}}
\item \href{#method-getDomain}{\code{TypeInst$getDomain()}}
\item \href{#method-setDomain}{\code{TypeInst$setDomain()}}
\item \href{#method-ranges}{\code{TypeInst$ranges()}}
\item \href{#method-isArray}{\code{TypeInst$isArray()}}
\item \href{#method-type}{\code{TypeInst$type()}}
\item \href{#method-clone}{\code{TypeInst$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
constuctor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TypeInst$new(type, indexExprVec = NULL, domain = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{type of declaration}

\item{\code{indexExprVec}}{expression list of indices}

\item{\code{domain}}{the domain of decision variables}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getDomain"></a>}}
\if{latex}{\out{\hypertarget{method-getDomain}{}}}
\subsection{Method \code{getDomain()}}{
get the variable domain
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TypeInst$getDomain()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setDomain"></a>}}
\if{latex}{\out{\hypertarget{method-setDomain}{}}}
\subsection{Method \code{setDomain()}}{
set the variable domain
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TypeInst$setDomain(dom)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dom}}{domain expression to be set}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ranges"></a>}}
\if{latex}{\out{\hypertarget{method-ranges}{}}}
\subsection{Method \code{ranges()}}{
return the index expression vector
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TypeInst$ranges()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-isArray"></a>}}
\if{latex}{\out{\hypertarget{method-isArray}{}}}
\subsection{Method \code{isArray()}}{
check if it's an array
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TypeInst$isArray()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-type"></a>}}
\if{latex}{\out{\hypertarget{method-type}{}}}
\subsection{Method \code{type()}}{
return the type information
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TypeInst$type()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TypeInst$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
