% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarDeclItem.R
\name{VarDeclItem}
\alias{VarDeclItem}
\title{The variable declaration item}
\description{
Declaration items in the model
}
\section{Super class}{
\code{\link[rminizinc:Item]{rminizinc::Item}} -> \code{VarDeclItem}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{.decl}}{the declaration expression}

\item{\code{.delete_flag}}{used to delete items}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{.decl}}{the declaration expression}

\item{\code{.delete_flag}}{used to delete items}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{VarDeclItem$new()}}
\item \href{#method-getDecl}{\code{VarDeclItem$getDecl()}}
\item \href{#method-setDecl}{\code{VarDeclItem$setDecl()}}
\item \href{#method-getId}{\code{VarDeclItem$getId()}}
\item \href{#method-c_str}{\code{VarDeclItem$c_str()}}
\item \href{#method-getDeleteFlag}{\code{VarDeclItem$getDeleteFlag()}}
\item \href{#method-delete}{\code{VarDeclItem$delete()}}
\item \href{#method-clone}{\code{VarDeclItem$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VarDeclItem$new(decl = NULL, mzn_str = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{decl}}{the declaration expression object}

\item{\code{mzn_str}}{string representation of variable declaration item}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getDecl"></a>}}
\if{latex}{\out{\hypertarget{method-getDecl}{}}}
\subsection{Method \code{getDecl()}}{
get the variable declaration
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VarDeclItem$getDecl()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setDecl"></a>}}
\if{latex}{\out{\hypertarget{method-setDecl}{}}}
\subsection{Method \code{setDecl()}}{
set the variable declaration
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VarDeclItem$setDecl(e)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{e}}{var decl expression}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getId"></a>}}
\if{latex}{\out{\hypertarget{method-getId}{}}}
\subsection{Method \code{getId()}}{
get the identifier object for the variable
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VarDeclItem$getId()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-c_str"></a>}}
\if{latex}{\out{\hypertarget{method-c_str}{}}}
\subsection{Method \code{c_str()}}{
set the variable declaration


convert the declaration to String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VarDeclItem$c_str()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getDeleteFlag"></a>}}
\if{latex}{\out{\hypertarget{method-getDeleteFlag}{}}}
\subsection{Method \code{getDeleteFlag()}}{
delete flag for internal use
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VarDeclItem$getDeleteFlag()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete"></a>}}
\if{latex}{\out{\hypertarget{method-delete}{}}}
\subsection{Method \code{delete()}}{
delete the variable item
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VarDeclItem$delete()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VarDeclItem$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
